Minor CSS Q.O.L. Updates & Bug Fixes
Updated overflow for bullet points to not go underneath bullet point, along with extra spacing between content and bullet point in order to improve readability.
Updated background of body to be same as background color in order to keep color consistent when refreshing page or dragging it around on mobile browsers.
Fixed " | " in between LinkedIn and GitHub not changing color when switching between dark and light mode.